Adding Unit tests

+++ hexl/include/hexl/thread-pool/thread-pool-vars.hpp
@@ -1,16 +1,17 @@
 // Copyright (C) 2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
+
 #pragma once
 
 namespace intel {
 namespace hexl {
 
 // Number of threads for thread pool
-extern int HEXL_NUM_THREADS;
+extern uint HEXL_NUM_THREADS;
 // Number of recursive levels as parallel calls
-extern int HEXL_NTT_PARALLEL_DEPTH;
+extern uint HEXL_NTT_PARALLEL_DEPTH;
 // Wait time (ms) before thread enters sleep mode
-extern uint64_t HEXL_THREAD_WAIT_TIME;
+extern uint16_t HEXL_THREAD_WAIT_TIME;
 
 }  // namespace hexl
 }  // namespace intel
