+++ hexl/thread-pool/thread-pool-vars.cpp
@@ -6,14 +6,17 @@
 namespace intel {
 namespace hexl {
 
-uint64_t HEXL_THREAD_WAIT_TIME = 20;  // ~ 100x wake up time on ICX machine
+uint16_t HEXL_THREAD_WAIT_TIME = 20;  // ~ 100x wake up time on ICX machine
+
+// const uint HEXL_DEFAULT_NUM_THREADS = 16;
+// const uint HEXL_DEFAULT_NTT_PARALLEL_DEPTH = 1;
 
 #ifdef HEXL_MULTI_THREADING
-int HEXL_NUM_THREADS = setup_num_threads("HEXL_NUM_THREADS");
-int HEXL_NTT_PARALLEL_DEPTH = setup_ntt_calls("HEXL_NTT_PARALLEL_DEPTH");
+uint HEXL_NUM_THREADS = setup_num_threads("HEXL_NUM_THREADS");
+uint HEXL_NTT_PARALLEL_DEPTH = setup_ntt_calls("HEXL_NTT_PARALLEL_DEPTH");
 #else
-int HEXL_NUM_THREADS = 1;
-int HEXL_NTT_PARALLEL_DEPTH = 0;
+uint HEXL_NUM_THREADS = 1;
+uint HEXL_NTT_PARALLEL_DEPTH = 0;
 #endif
 
 }  // namespace hexl
